Develop

+++ src/Controller/TaskController.php
@@ -3,10 +3,9 @@
 namespace App\Controller;
 
 use App\Entity\Task;
-use App\Entity\User;
 use App\Form\TaskType;
 use App\Repository\TaskRepository;
-use Doctrine\ORM\EntityManager;
+use App\Service\SecurityService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
@@ -16,18 +15,21 @@ use Symfony\Component\Routing\Attribute\Route;
 
 class TaskController extends AbstractController
 {
-    #[Route("/", name:"task_list")]
+    public function __construct(private readonly SecurityService $securityService)
+    {
+    }
+    #[Route("/", name: "task_list")]
     public function listAction(TaskRepository $taskRepository): Response
     {
         $currentUser = $this->getUser();
-        if(in_array('ROLE_ADMIN', $currentUser->getRoles())) {
+        if ($currentUser !== null && in_array('ROLE_ADMIN', $currentUser->getRoles())) {
             return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findAll()]);
         }
 
-        return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findBy([ 'user' => $currentUser])]);
+        return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findBy(['user' => $currentUser])]);
     }
 
-    #[Route("/tasks/create", name:"task_create")]
+    #[Route("/tasks/create", name: "task_create")]
     public function createAction(Request $request, EntityManagerInterface $entityManager): RedirectResponse|Response
     {
         $task = new Task();
@@ -47,11 +49,16 @@ class TaskController extends AbstractController
         return $this->render('task/create.html.twig', ['form' => $form->createView()]);
     }
 
-    #[Route("/tasks/{id}/edit", name:"task_edit")]
+    #[Route("/tasks/{id}/edit", name: "task_edit")]
     public function editAction(Task $task, Request $request, EntityManagerInterface $entityManager): RedirectResponse|Response
     {
-        $form = $this->createForm(TaskType::class, $task);
+        if (!$this->securityService->checkUpPermission($task)) {
+            $this->addFlash('error', 'Vous n\'êtes pas autorisé à modifier cette tâche.');
 
+            return $this->redirectToRoute('task_list');
+        }
+
+        $form = $this->createForm(TaskType::class, $task);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -68,9 +75,15 @@ class TaskController extends AbstractController
         ]);
     }
 
-    #[Route("/tasks/{id}/toggle", name:"task_toggle")]
+    #[Route("/tasks/{id}/toggle", name: "task_toggle")]
     public function toggleTaskAction(Task $task, EntityManagerInterface $entityManager): RedirectResponse
     {
+        if (!$this->securityService->checkUpPermission($task)) {
+            $this->addFlash('error', 'Vous n\'êtes pas autorisé à modifier cette tâche.');
+
+            return $this->redirectToRoute('task_list');
+        }
+
         $task->toggle(!$task->isDone());
         $entityManager->flush();
 
@@ -79,9 +92,15 @@ class TaskController extends AbstractController
         return $this->redirectToRoute('task_list');
     }
 
-    #[Route("/tasks/{id}/delete", name:"task_delete")]
+    #[Route("/tasks/{id}/delete", name: "task_delete")]
     public function deleteTaskAction(Task $task, EntityManagerInterface $entityManager): RedirectResponse
     {
+        if (!$this->securityService->checkUpPermission($task)) {
+            $this->addFlash('error', 'Vous n\'êtes pas autorisé à modifier cette tâche.');
+
+            return $this->redirectToRoute('task_list');
+        }
+
         $entityManager->remove($task);
         $entityManager->flush();
